refactor: reorganize tests and improve documentation
- Added a doc comment to the `download` function in `feed.rs` to clarify its purpose.
- Moved the `verify_accepts_good_signature_and_rejects_tampering` test from `install.rs` to its own test module, improving test organization.
- Restored the `maps_special_and_all_function_keys` test in `keys.rs` to ensure coverage of key mapping functionality.
- Removed redundant test cases from `recorder.rs` to streamline the test suite.
