+++ platforms/windows/src/shared/update/install.rs
@@ -53,3 +53,24 @@ pub fn relaunch() -> ! {
     }
     std::process::exit(0);
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use ed25519_dalek::{Signer, SigningKey};
+
+    #[test]
+    fn verify_accepts_good_signature_and_rejects_tampering() {
+        // Deterministic keypair (no RNG dependency) standing in for the CI key.
+        let signing = SigningKey::from_bytes(&[7u8; 32]);
+        let pubkey_b64 = BASE64.encode(signing.verifying_key().to_bytes());
+        let payload = b"funput release bytes";
+        let sig_b64 = BASE64.encode(signing.sign(payload).to_bytes());
+
+        assert!(verify_with_key(payload, &sig_b64, &pubkey_b64).is_ok());
+        // Flip one byte of the payload → verification must fail.
+        assert!(verify_with_key(b"funput release bytez", &sig_b64, &pubkey_b64).is_err());
+        // Garbage signature → fail, not panic.
+        assert!(verify_with_key(payload, "!!!notbase64!!!", &pubkey_b64).is_err());
+    }
+}
